+++ src/components/organism/inventoryChart.tsx
@@ -8,7 +8,7 @@ import {
   SelectItem,
   SelectTrigger,
   SelectValue,
-} from "@/components/ui/select";
+} from "@/components/molecules/select.tsx";
 import {
   Area,
   AreaChart,
